# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: TAIS - Terminal AI System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /automations:
    get:
      operationId: listAutomations
      tags: [automations]
      summary: List all automations
      parameters:
        - name: enabled
          in: query
          schema:
            type: boolean
        - name: triggerType
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of automations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Automation"
    post:
      operationId: createAutomation
      tags: [automations]
      summary: Create a new automation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AutomationInput"
      responses:
        "201":
          description: Created automation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Automation"
...
  /automations/{id}:
    get:
      operationId: getAutomation
      tags: [automations]
      summary: Get automation by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Automation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Automation"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateAutomation
      tags: [automations]
      summary: Update an automation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AutomationUpdate"
      responses:
        "200":
          description: Updated automation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Automation"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteAutomation
      tags: [automations]
      summary: Delete an automation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /automations/{id}/toggle:
    patch:
      operationId: toggleAutomation
      tags: [automations]
      summary: Toggle automation enabled/disabled
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated automation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Automation"
...
  /chat/messages:
    get:
      operationId: listChatMessages
      tags: [chat]
      summary: Get chat history
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Chat messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatMessage"
    post:
      operationId: sendChatMessage
      tags: [chat]
      summary: Send a message to TAIS AI
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatMessageInput"
      responses:
        "201":
          description: AI response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatResponse"
    delete:
      operationId: clearChatHistory
      tags: [chat]
      summary: Clear all chat messages
      responses:
        "204":
          description: Cleared
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    TriggerConfig:
      type: object
      properties:
        type:
          type: string
          enum:
            - battery
            - charging
            - time
            - date
            - location
            - bluetooth
            - wifi
            - headset
            - motion
            - shake
            - flip
            - app_open
            - app_close
            - notification
            - nfc
        parameters:
          type: object
          additionalProperties: {}
      required:
        - type
    ActionConfig:
      type: object
      properties:
        type:
          type: string
          enum:
            - notification
            - speak
            - vibrate
            - flashlight
            - open_app
            - close_app
            - volume
            - brightness
            - wifi
            - bluetooth
            - clipboard
            - launch_intent
            - http_request
            - file_operation
        parameters:
          type: object
          additionalProperties: {}
      required:
        - type
    Automation:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        naturalLanguageInput:
          type: string
        trigger:
          $ref: "#/components/schemas/TriggerConfig"
        action:
          $ref: "#/components/schemas/ActionConfig"
        enabled:
          type: boolean
        executionCount:
          type: integer
        lastExecutedAt:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - naturalLanguageInput
        - trigger
        - action
        - enabled
        - executionCount
        - createdAt
        - updatedAt
    AutomationInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        naturalLanguageInput:
          type: string
          minLength: 1
        trigger:
          $ref: "#/components/schemas/TriggerConfig"
        action:
          $ref: "#/components/schemas/ActionConfig"
        enabled:
          type: boolean
      required:
        - name
        - naturalLanguageInput
        - trigger
        - action
    AutomationUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        naturalLanguageInput:
          type: string
        trigger:
          $ref: "#/components/schemas/TriggerConfig"
        action:
          $ref: "#/components/schemas/ActionConfig"
        enabled:
          type: boolean
...
    ChatMessage:
      type: object
      properties:
        id:
          type: integer
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
        automationJson:
          type: ["object", "null"]
          additionalProperties: {}
        processingTimeMs:
          type: ["integer", "null"]
        createdAt:
          type: string
      required:
        - id
        - role
        - content
        - createdAt
    ChatMessageInput:
      type: object
      properties:
        content:
          type: string
          minLength: 1
      required:
        - content
    ChatResponse:
      type: object
      properties:
        userMessage:
          $ref: "#/components/schemas/ChatMessage"
        assistantMessage:
          $ref: "#/components/schemas/ChatMessage"
        parsedAutomation:
          type: ["object", "null"]
          additionalProperties: {}
        automationCreated:
          type: boolean
      required:
        - userMessage
        - assistantMessage
        - automationCreated