# lib/api-zod/src/generated/types/automation.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * TAIS - Terminal AI System API
 * OpenAPI spec version: 1.0.0
 */
import type { ActionConfig } from './actionConfig';
import type { TriggerConfig } from './triggerConfig';

export interface Automation {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  naturalLanguageInput: string;
  trigger: TriggerConfig;
  action: ActionConfig;
  enabled: boolean;
  executionCount: number;
  /** @nullable */
  lastExecutedAt?: string | null;
  createdAt: string;
  updatedAt: string;
}