# lib/api-zod/src/generated/types/automationInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * TAIS - Terminal AI System API
 * OpenAPI spec version: 1.0.0
 */
import type { ActionConfig } from './actionConfig';
import type { TriggerConfig } from './triggerConfig';

export interface AutomationInput {
  /** @minLength 1 */
  name: string;
  description?: string;
  /** @minLength 1 */
  naturalLanguageInput: string;
  trigger: TriggerConfig;
  action: ActionConfig;
  enabled?: boolean;
}