# lib/api-zod/src/generated/types/automationUpdate.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * TAIS - Terminal AI System API
 * OpenAPI spec version: 1.0.0
 */
import type { ActionConfig } from './actionConfig';
import type { TriggerConfig } from './triggerConfig';

export interface AutomationUpdate {
  /** @minLength 1 */
  name?: string;
  description?: string;
  naturalLanguageInput?: string;
  trigger?: TriggerConfig;
  action?: ActionConfig;
  enabled?: boolean;
}