# lib/api-zod/src/generated/types/triggerConfig.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * TAIS - Terminal AI System API
 * OpenAPI spec version: 1.0.0
 */
import type { TriggerConfigParameters } from './triggerConfigParameters';
import type { TriggerConfigType } from './triggerConfigType';

export interface TriggerConfig {
  type: TriggerConfigType;
  parameters?: TriggerConfigParameters;
}